+++ packages/core/src/lib/graticule.ts
@@ -1,6 +1,7 @@
 import type { GeoGraticuleGenerator } from 'd3-geo'
 
 import type { MapContext } from './map'
+import type { MapObjectProps } from './mapObject'
 import type { MethodsToModifiers } from './utils'
 
 import { geoGraticule } from 'd3-geo'
@@ -16,6 +17,15 @@ import { applyModifiers } from './utils'
  * @see https://d3js.org/d3-geo/shape#geoGraticule
  */
 export interface GraticuleConfig extends MethodsToModifiers<GeoGraticuleGenerator> {}
+
+/**
+ * Shared props contract for graticule layers.
+ */
+export interface MapGraticuleProps<TStyle = unknown> extends MapObjectProps<TStyle> {
+  config?: GraticuleConfig
+  background?: boolean | string
+  border?: boolean | string
+}
 
 export function renderGraticule(context: MapContext, config?: GraticuleConfig): string | null {
   const graticule = geoGraticule()