+++ packages/react/src/hooks/useCreateMapContext.ts
@@ -1,22 +1,22 @@
 'use client'
 
 import type {
-  MapConfig,
   MapContext,
+  MapProps,
 } from '@d3-maps/core'
 
 import { makeMapContext } from '@d3-maps/core'
 import { useMemo } from 'react'
 
 export function useCreateMapContext(
-  config?: Partial<MapConfig>,
+  config?: Partial<MapProps>,
   context?: MapContext,
 ): MapContext | undefined {
   return useMemo(() => {
     if (context) return context
     if (!config?.data) return undefined
 
-    return makeMapContext(config as MapConfig)
+    return makeMapContext(config as MapProps)
   }, [
     context,
     config?.width,