+++ packages/react/tests/mapZoom.spec.tsx
@@ -5,7 +5,7 @@ import {
   screen,
 } from '@testing-library/react'
 
-import { Map } from '../src/components/Map'
+import { MapBase } from '../src/components/MapBase'
 import { MapZoom } from '../src/components/MapZoom'
 import { sampleGeoJson } from './fixtures'
 
@@ -46,7 +46,7 @@ describe('mapZoom', () => {
     const transition = { duration: 250 }
 
     const { rerender } = render(
-      <Map data={sampleGeoJson}>
+      <MapBase data={sampleGeoJson}>
         <MapZoom
           data-testid="map-zoom-group"
           center={[11, 12]}
@@ -61,7 +61,7 @@ describe('mapZoom', () => {
         >
           <g data-testid="zoom-content" />
         </MapZoom>
-      </Map>,
+      </MapBase>,
     )
 
     expect(setupZoomSpy).toHaveBeenCalled()
@@ -82,7 +82,7 @@ describe('mapZoom', () => {
     expect(zoomGroup?.getAttribute('transform')).toContain('translate')
 
     rerender(
-      <Map data={sampleGeoJson}>
+      <MapBase data={sampleGeoJson}>
         <MapZoom
           data-testid="map-zoom-group"
           center={[99, 100]}
@@ -97,7 +97,7 @@ describe('mapZoom', () => {
         >
           <g data-testid="zoom-content" />
         </MapZoom>
-      </Map>,
+      </MapBase>,
     )
 
     expect(applyZoomSpy).toHaveBeenCalledWith(