+++ packages/vue/src/components/index.ts
@@ -1,4 +1,4 @@
-import Map from './Map.vue'
+import MapBase from './MapBase.vue'
 import MapFeature from './MapFeature.vue'
 import MapFeatures from './MapFeatures.vue'
 import MapGraticule from './MapGraticule.vue'
@@ -7,7 +7,7 @@ import MapMesh from './MapMesh.vue'
 import MapZoom from './MapZoom.vue'
 
 export {
-  Map,
+  MapBase,
   MapFeature,
   MapFeatures,
   MapGraticule,