+++ packages/vue/tests/mapMarker.spec.ts
@@ -4,14 +4,14 @@ import { mount } from '@vue/test-utils'
 import { h } from 'vue'
 
 import {
-  Map,
+  MapBase,
   MapMarker,
 } from '../src'
 import { sampleGeoJson } from './fixtures'
 
 describe('mapMarker', () => {
   it('uses projection transform from context', () => {
-    const wrapper = mount(Map, {
+    const wrapper = mount(MapBase, {
       props: {
         data: sampleGeoJson,
         width: 400,
@@ -38,7 +38,7 @@ describe('mapMarker', () => {
   })
 
   it('recomputes marker transform when map context changes', async () => {
-    const wrapper = mount(Map, {
+    const wrapper = mount(MapBase, {
       props: {
         data: sampleGeoJson,
         width: 300,